feat: refactor analytics files and add new event
